client/requirements.txt
psutil
requests
python-dotenv
py-cpuinfo
